Moved most dag operations from makeflow_commong to dag.c.
