+++ makeflow/src/makeflow_analyze.c
@@ -45,7 +45,8 @@ See the file COPYING for details.
 #include "path.h"
 
 #include "dag.h"
-#include "visitors.h"
+#include "dag_visitors.h"
+#include "parser.h"
 
 #include "makeflow_common.h"
