+++ makeflow/src/makeflow_viz.c
@@ -44,7 +44,8 @@ See the file COPYING for details.
 #include "path.h"
 
 #include "dag.h"
-#include "visitors.h"
+#include "dag_visitors.h"
+#include "parser.h"
 
 #include "makeflow_common.h"
