+++ makeflow/src/parser.h
@@ -1,0 +1,12 @@
+/*
+Copyright (C) 2008- The University of Notre Dame
+This software is distributed under the GNU General Public License.
+See the file COPYING for details.
+*/
+
+#ifndef PARSER_H
+#define PARSER_H
+
+struct dag *dag_from_file(const char *filename);
+
+#endif
